cookbook/starter-apps/Jee_problem_solver_and_analyzer/requirements.txt
streamlit
openai
educhain
Pillow
langchain_openai